drivers: sensor: tmag5273: 1-byte read mode
TMAG5273 (and TMAG3001) supports a "1-byte read mode" where an I2C read
issued without a prior I2C write to specify read address can be used to
read the axis measurements in either 8 or 16 bit resolution. This allows
for faster reading of multiple sensors by reducing amount of I2C chatter.

This commit adds support for this read mode and allows it to be configured
in the DT. Currently, it only supports CONTINUOUS sampling mode, as I
couldn't get manual triggering in standby mode to work while this read mode
is enabled.

See datasheet:
- TMAG5273 "6.5.1.3.4 1-Byte I2C Read Command for 16-Bit Data"
- TMAG3001 "6.5.1.2.4 I2C Read Command for 16-bit Data"

## Test plan

Tested in continuous mode on TMAG5273 with single axis and all 3 axis. Also
tested with crc checking. I do not have a TMAG3001 to test its behavior
there.

Signed-off-by: Yiding Jia <yiding.jia@gmail.com>